package/python-pyqt5-sip: bump version to 12.15.0
The versioning scheme of this package is not trivial: the PyQt5-sip
version number does not corresponds to the actual version tag name in
its Python-SIP source repository. For that reason, this commit adds a
comment explaining this peculiar mapping. This comment will help, in
future maintenance tasks.

To better understand the upstream changes in this package explained in
this commit log, it is worth mentioning that:
- PyQt5-sip 12.13.0 was generated with Python-SIP 6.7.12, and
- PyQt5-sip 12.15.0 was generated with Python-SIP 6.8.6.

The package license has changed to BSD-2-Clause, in upstream commits:
https://github.com/Python-SIP/sip/commit/0ba3b233ec8ed7f22f2f2f950b64f61b50ddfc00
https://github.com/Python-SIP/sip/commit/4dde1491f7b8ace15833cffb86bee82d60683968
https://github.com/Python-SIP/sip/commit/f32039b07c47280039dcdcd0f8a2b1323fda912e

This commit updates the _LICENSE and _LICENSE_FILES macro
accordingly. The license hash file is also updated. While at it, this
commit also adds the MD5 hash published upstream.

This commit also introduces minor improvements to better follow
Buildroot package conventions:
- the .mk file header is changed to lower case,
- hash file comments are updated,
- the locally calculated license file hash is changed from MD5 to
  SHA256.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>