configs/qemu_aarch64_sbsa_defconfig: update ATF to v2.11
qemu_aarch64_sbsa_defconfig doesn't boot since the qemu version bump to
9.0 [1]. Sync the ATF version with the one (v2.11) used by Qemu tests
[2].

Since we use a released version of ATF, we switch to ATF custom version
to use the official ATF repository instead of the read-only mirror on
ARM-software's github.

This doesn't fix yet this defconfig due to a bug in edk2 as mentioned
in the qemu commit log [2].

With the edk2 version bump applied, this defconfig doesn't boot with
ATF v2.10-lts but boot successfully with ATF v2.11.

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/103a55028c75dace312470ba05a197886c6b6100
[2] https://gitlab.com/qemu-project/qemu/-/commit/24a7cd6a7c21f82f1ce9bd5ecf6fb54eb7bf4602

Signed-off-by: Romain Naour <romain.naour@smile.fr>
Signed-off-by: Julien Olivain <ju.o@free.fr>