support/testing: test_grub: update and re-enable the test
The EDK2 bump to stable202405 introduced an incompatibility with the
Qemu version 5.2.0 present in the reference Docker image. For that
reason, this test was disabled.

This commit enables host-qemu in the build, to be used by the runtime
test infrastructure.

Since this commit switch to a more recent host-qemu, it also removes
firmware options to workaround old qemu bugs.

Those changes are fixing this test, which is why this commit re-enables
it by removing the skipTest() invocation.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Reviewed-by: Romain Naour <romain.naour@smile.fr>
Signed-off-by: Romain Naour <romain.naour@smile.fr>