package/bsdiff: drop package
As advocated by Thomas Petazzoni [1], drop bsdiff as it is not maintained
anymore.

Meanwhile, autobuild also detected that the package was no longer
downloadable from the URL, and indeed, Colin Percival, the maintainer,
confirmed that he intentionally disabled the package download due to the
presence of some security bugs that he doesn't have time to fix.

[1] https://patchwork.ozlabs.org/project/buildroot/patch/20241004215451.1198861-2-dario.binacchi@amarulasolutions.com/

Fixes:
- http://autobuild.buildroot.org/results/21bb827d3dbf553c79f49d7c184099c75df3d45d

Suggested-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>