package/findutils: bump version to 4.10.0
For release announce, see:
https://lists.gnu.org/archive/html/info-gnu/2024-07/msg00002.html

This commit also updates the hash file comment, to use the sha256 hash
published in the release announce, to provide the signature URL and the
pgp key id used for the signature verification.

See the gnulib documentation for details [1]. Contrary to what the
option name might suggest, it doesn't really disable Y2038 support,
but only the check that the system is Y2038 compliant. So even with
--disable-year2038, if the system is Y2038 compliant (uses a 64-bit
arch, uses the musl C library, or uses the glibc C library with
BR2_TIME_BITS_64=y), findutils will be Y2038 compliant.

[1] https://www.gnu.org/software/gnulib/manual/html_node/Avoiding-the-year-2038-problem.html

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>