package/python3: select host Python 3
Building Python for the target requires host Python via
PYTHON3_DEPENDENCIES. Host Python is exposed as a standalone Kconfig
option, so when BR2_PACKAGE_PYTHON3=y, it is possible for host Python to
be built even when BR2_PACKAGE_HOST_PYTHON3=n. Normally this is OK for
host packages. However, host-python3 has a number of sub-options, which
sometimes need to be selected by a package. In that case, that
particular package has to add the `select BR2_PACKAGE_HOST_PYTHON3`
while it doesn't have a direct dependency on host-python3.

Add `select BR2_PACKAGE_HOST_PYTHON3' to python3 itself.

Signed-off-by: Liam Hupfer <lhupfer@xes-inc.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>