package/bc: update _SITE to use BR2_GNU_MIRROR
The bc package is distributed on the GNU project servers. See [1]
and [2]. Buildroot has the BR2_GNU_MIRROR configuration which can
be used for that purpose. See [3]

For consistency with all other GNU packages, this commit updates
the _SITE to use BR2_GNU_MIRROR.

As a side note, the bc package was introduced long time ago using
BR2_GNU_MIRROR. See [4]. It was then updated to an alpha version
in [5]. When the alpha version was no longer needed, it was switched
to the main GNU download server in [6].

[1] https://www.gnu.org/software/bc/
[2] https://www.gnu.org/prep/ftp.html
[3] https://gitlab.com/buildroot.org/buildroot/-/blob/2024.08.2/Config.in#L286
[4] https://gitlab.com/buildroot.org/buildroot/-/commit/c95dcd464512ede5ffdef5567558611f05e2e52b
[5] https://gitlab.com/buildroot.org/buildroot/-/commit/1faa7c344e10986a4afd1d17bf6d4265b3c821fc
[6] https://gitlab.com/buildroot.org/buildroot/-/commit/ed7572cc7ff69584116c4e763a608afc225bd0c6

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Romain Naour <romain.naour@smile.fr>
(cherry picked from commit 2956a3921548921adb5472e092901ff9d9861333)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>