configs/imx8mn_bsh_smm_s2[_pro]: bump U-Boot to 2024.10 and Linux to 6.11.8
The patch bumps U-Boot to version 2024.10 and the Linux kernel to version
6.11.8.
The new version of U-Boot has added EFI Capsule support, which depends on
GnuTLS, thus requiring the activation of BR2_TARGET_UBOOT_NEEDS_GNUTLS.
U-Boot also needs util-linux, therefore
BR2_TARGET_UBOOT_NEEDS_UTIL_LINUX is added.
The version of TF-A remains 2.10, but the LTS version is used as suggested
by the maintainers.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
[Julien: added BR2_TARGET_UBOOT_NEEDS_UTIL_LINUX=y]
Signed-off-by: Julien Olivain <ju.o@free.fr>