configs/versal_vek280_defconfig: bump to xilinx-v2024.2
This patch bumps the versal_vek280_defconfig to xilinx-v2024.2 which includes
the following updates:

- ATF v2.10
- Linux v6.6.40
- PLM xilinx-v2024.2
- PSMFW xilinx-v2024.2
- U-Boot v2024.01

The board specific Linux patch for the vek280 needs to be changed for the
2024.2 release.  While the Linux vek280 dts files were correctly updated, the
prebuilt vpl_gen_fixed.pdi file is still not defining the DDR memory region
from address 0x8 0000 0000 to 0x8 8000 0000.

For this reason, a new patch is needed for 2024.2 which is only a work around
until the pdi file is corrected.  This is why the dts patch will not be
applied upstream.

The following patch has thus been removed:
board/versal/vek280/patches/linux/0001-arm64-versal-Fix-vek280-memory-configuration.patch

And replaced with:
board/versal/vek280/patches/linux/0001-arm64-versal-vek280-prebuilt-pdi-is-incorrect.patch

Upstream: n/a

Signed-off-by: Neal Frager <neal.frager@amd.com>
[Julien: remove the package patch numbering]
Signed-off-by: Julien Olivain <ju.o@free.fr>