package/libpam-radius-auth: bump to version 3.0.0
For change log, see [1].

Buildroot commit [2] changed _SITE from using the ftp site to a
github download. This github generated archive has a different hash
compare to release files, due to a different directory name in the
archive. This does not allow to check the published PGP signatures.
This new 3.0.0 version is published as a github release asset. This
commit updates _SITE and _SOURCE to use those release files. The
archive is switched to bz2 archive which is smaller. Finally, this
commit adds a comment in the hash file, with the signature link and
the pgp key used for the signature.

[1] https://github.com/FreeRADIUS/pam_radius/blob/release_3_0_0/Changelog
[2] https://gitlab.com/buildroot.org/buildroot/-/commit/860b1b958797bd16bfe7666bb308a5ea026b1b46

Signed-off-by: Vincent Jardin <vjardin@free.fr>
[Julien: change to use the release archive so the PGP
  signature can be checked]
Signed-off-by: Julien Olivain <ju.o@free.fr>