package/android-tools: convert all patches to Git-formatted patches
This will make it easier to modify the patch series moving forward. No
functional changes.

On one patch, we added the SoB from Giulio, which was missing. Giulio
was the original contributor of the patch, and nobody touched it since
it was introduced.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>
(cherry picked from commit 3ee745782a0802b1479eb6a8f678b6a261775295)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>