package/linuxptp: bump version to 4.4
The released package is no longer available under sourceforge. So we
switch the download URL to the official download page [1]. Also the info URL
about the package source is changed to the official LinuxPTP project page [2].

The release note [3] and annoucement on the mailinglist [4] give more
information about the changes.

[1] http://downloads.nwtime.org/linuxptp/
[2] https://linuxptp.nwtime.org/about/features/
[3] https://www.nwtime.org/news/linuxptp-4.4-released/
[4] https://lists.nwtime.org/sympa/arc/linuxptp-devel/2024-09/msg00020.html

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>
[Julien:
  - switch _SITE to https
  - keep the original _SOURCE using _VERSION
]
Signed-off-by: Julien Olivain <ju.o@free.fr>