configs/rock5b: Switch to mainline TrustedFirmware
The Rock 5b board support currently uses a binary trusted firmware
blob from the rockchip-rkbin package. Recently, the Rock 5b's SoC,
i.e., the rk3588, received mainline TrustedFirmware-A support:

https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/21840
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/29363

This support was recently released as part of the v2.12 version of
the TrustedFirmware-A. This patch changes the defconfig from using
the rockchip-rkbin atf binary blob to building the TrustedFirmware-A
v2.12 for the rk3588 from source.

Signed-off-by: Kilian Zinnecker <kilian.zinnecker@mail.de>
Signed-off-by: Julien Olivain <ju.o@free.fr>