package/libcamera: bump to version 0.4.0
For change log since 0.3.2, see:
https://git.libcamera.org/libcamera/libcamera.git/tag/?h=v0.4.0

libcamera dropped the gcc 8 support in [1] and increased the minimal
version to 9. This commit updates that requirement in the package
Config.in and in reverse dependency (in libcamera-apps).

This commit also updates the libcamera runtime test to the new format
of the "cam --list-properties" output.

[1] https://git.libcamera.org/libcamera/libcamera.git/commit/?id=9f513439b6896d1a33c91a400bd02b7b8829feb6

Cc: Kieran Bingham <kieran.bingham@ideasonboard.com>
Cc: Marcus Folkesson <marcus.folkesson@gmail.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>