package/flashrom: bump to version 1.5.1
Changes in 1.5.0 and 1.5.1 are described at:
https://www.flashrom.org/release_notes/v_1_5.html

The option classic_cli_print_wiki has been removed in upstream commit
87134f538dd9b69fa1331514f7b492e92f8bd66d ("tree: Remove
print_wiki.c"). It was used to generate some parts of the Wiki, and
was anyway not used by Buildroot as it was disabled.

The build was tested including with fwupd, which is the one reverse
dependency of flashrom in Buildroot.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>